+++ Gravitas.Platform.Web/ViewModel/MixedFeedManage/03_AddOperationVisaVm.cs
@@ -1,9 +1,10 @@
-﻿namespace Gravitas.Platform.Web.ViewModel {
-
-	public static partial class MixedFeedManageVms
-	{
-		public class AddOperationVisaVm {
-			public long NodeId { get; set; }
-		}
-	}
+﻿namespace Gravitas.Platform.Web.ViewModel
+{
+    public static partial class MixedFeedManageVms
+    {
+        public class AddOperationVisaVm
+        {
+            public int NodeId { get; set; }
+        }
+    }
 }
