+++ Gravitas.Platform.Web/ViewModel/OpRoutine/CentralLabolatoryProcess/06_PrintAddOpVisaVm.cs
@@ -4,7 +4,7 @@
     {
         public class PrintAddOpVisaVm
         {
-            public long NodeId { get; set; }
+            public int NodeId { get; set; }
         }
     }
 }
