+++ Gravitas.Platform.Web/ViewModel/OpRoutine/LabolatoryIn/27_PrintLaboratoryProtocol.cs
@@ -1,12 +1,13 @@
 ﻿using System;
 
-namespace Gravitas.Platform.Web.ViewModel {
-
-	public static partial class LaboratoryInVms {
-
-		public class PrintLaboratoryProtocol {
-			public long NodeId { get; set; }
-			public Guid OpDataId { get; set; }
-		}
-	}
+namespace Gravitas.Platform.Web.ViewModel
+{
+    public static partial class LaboratoryInVms
+    {
+        public class PrintLaboratoryProtocol
+        {
+            public int NodeId { get; set; }
+            public Guid OpDataId { get; set; }
+        }
+    }
 }
