+++ Gravitas.Platform.Web/ViewModel/OpRoutine/LoadCheckPoint/01_IdleVm.cs
@@ -1,12 +1,13 @@
 ﻿using System;
 
-namespace Gravitas.Platform.Web.ViewModel {
-
-	public static partial class LoadCheckPointVms {
-
-		public class IdleVm {
-			public long NodeId { get; set; }
-			public LoadPointTicketContainerItemVm BindedTruck { get; set; }
-		}
-	}
+namespace Gravitas.Platform.Web.ViewModel
+{
+    public static partial class LoadCheckPointVms
+    {
+        public class IdleVm
+        {
+            public int NodeId { get; set; }
+            public LoadPointTicketContainerItemVm BindedTruck { get; set; }
+        }
+    }
 }
