+++ Gravitas.Platform.Web/ViewModel/OpRoutine/SingleWindow/04. ContainerCloseAddOpVisaVm.cs
@@ -1,9 +1,10 @@
-﻿namespace Gravitas.Platform.Web.ViewModel {
-
-	public static partial class SingleWindowVms {
-
-		public class ContainerCloseAddOpVisaVm {
-			public long NodeId { get; set; }
-		}
-	}
+﻿namespace Gravitas.Platform.Web.ViewModel
+{
+    public static partial class SingleWindowVms
+    {
+        public class ContainerCloseAddOpVisaVm
+        {
+            public int NodeId { get; set; }
+        }
+    }
 }
