+++ Gravitas.Platform.Web/ViewModel/OpRoutine/SingleWindow/14. EditPostApiData.cs
@@ -1,14 +1,15 @@
 ﻿using System;
 
-namespace Gravitas.Platform.Web.ViewModel {
-
-	public static partial class SingleWindowVms {
-
-		public class EditPostApiDataVm {
-			public long NodeId { get; set; }
-			public long OpDataId { get; set; }
-			public string SupplyCode { get; set; }
-			public DateTime RequestSentDateTime { get; set; }
-		}
-	}
+namespace Gravitas.Platform.Web.ViewModel
+{
+    public static partial class SingleWindowVms
+    {
+        public class EditPostApiDataVm
+        {
+            public int NodeId { get; set; }
+            public int OpDataId { get; set; }
+            public string SupplyCode { get; set; }
+            public DateTime RequestSentDateTime { get; set; }
+        }
+    }
 }
