+++ Gravitas.Platform.Web/ViewModel/OpRoutine/Weightbridge/15_GetTrailerWeightVm.cs
@@ -1,11 +1,10 @@
 ﻿namespace Gravitas.Platform.Web.ViewModel
 {
-
     public static partial class WeightbridgeVms
     {
         public class GetTrailerWeightVm
         {
-            public long NodeId { get; set; }
+            public int NodeId { get; set; }
         }
     }
 }
