+++ Gravitas.Platform.Web/ViewModel/OpRoutine/Weightbridge/17_OpenBarrierOutVm.cs
@@ -1,11 +1,10 @@
 ﻿namespace Gravitas.Platform.Web.ViewModel
 {
-
     public static partial class WeightbridgeVms
     {
         public class OpenBarrierOutVm
         {
-            public long NodeId { get; set; }
+            public int NodeId { get; set; }
             public bool IsSmsNecessary { get; set; }
         }
     }
