+++ Gravitas.Platform.Web/ViewModel/Ticket/History/TicketHistoryItem.cs
@@ -4,8 +4,8 @@ namespace Gravitas.Platform.Web.ViewModel
 {
     public class TicketHistoryItem
     {
-        public long TicketId { get; set; }
-        public long TicketContainerId { get; set; }
+        public int TicketId { get; set; }
+        public int TicketContainerId { get; set; }
         [DisplayName("Картка")]
         public string CardNumber { get; set; }
         [DisplayName("Транспорт")]
