+++ Gravitas.Platform.Web/Views/Filters/OpDataRegistryFilters.cs
@@ -5,13 +5,13 @@ namespace Gravitas.Platform.Web.Models
 {
     public class OpDataRegistryFilters
     {
-        public long? TicketContainerId { get; set; }
-        public long? TicketId { get; set; }
+        public int? TicketContainerId { get; set; }
+        public int? TicketId { get; set; }
         public IEnumerable<long> NodeIds { get; set; }
-        public long? NodeType { get; set; }
-        public long? OpDataType { get; set; }
+        public int? NodeType { get; set; }
+        public int? OpDataType { get; set; }
         public DateTime? ArrivalDateTime { get; set; }
         public DateTime? DepartureDateTime { get; set; }
-        public long? StateId { get; set; }
+        public int? StateId { get; set; }
     }
 }
